Refcator the navbar
